Merge down changes for 0.1.3

+++ lib/openstudio_measure_tester/openstudio_testing_result.rb
@@ -52,8 +52,8 @@ module OpenStudioMeasureTester
         if !logs.empty?
           sha = logs.first.sha
         end
-      rescue StandardError => error 
-        puts "Could not find .git for measure(s), will not be able to report git information"
+      rescue StandardError => error
+        puts 'Could not find .git for measure(s), will not be able to report git information'
       end
 
       # check if the results data already exist, and if so, then load the file now to keep the results
@@ -76,6 +76,7 @@ module OpenStudioMeasureTester
           FileUtils.mv "#{@results_dir}/coverage", "#{@test_results_dir}/."
 
           cov = OpenStudioMeasureTester::Coverage.new("#{@test_results_dir}/coverage")
+          cov.parse_results
           @results['coverage'] = cov.to_hash
         end
 
@@ -83,8 +84,14 @@ module OpenStudioMeasureTester
         if Dir.exist?("#{@results_dir}/test/html_reports") || Dir.exist?("#{@results_dir}/test/reports")
           FileUtils.rm_rf "#{@test_results_dir}/minitest" if Dir.exist? "#{@test_results_dir}/minitest"
           FileUtils.mkdir_p "#{@test_results_dir}/minitest"
-          FileUtils.mv "#{@results_dir}/test/html_reports", "#{@test_results_dir}/minitest/html_reports"
-          FileUtils.mv "#{@results_dir}/test/reports", "#{@test_results_dir}/minitest/reports"
+
+          if Dir.exist?("#{@results_dir}/test/html_reports")
+            FileUtils.mv "#{@results_dir}/test/html_reports", "#{@test_results_dir}/minitest/html_reports"
+          end
+
+          if Dir.exist?("#{@results_dir}/test/reports")
+            FileUtils.mv "#{@results_dir}/test/reports", "#{@test_results_dir}/minitest/reports"
+          end
 
           # Delete the test folder if it is empty
           FileUtils.rm_rf "#{@results_dir}/test" if Dir.entries("#{@results_dir}/test").size == 2
